delete manualmode's jbossas-messaging-failover before copy
HornetQ failover is depending on the file system to lock a file.
I have seen PRs be aborted because HornetQ live server can not lock this
file to start (there is no timeout on this lock operation).
Deleting the directory ensures that the file system's lock (from previous
runs) is removed.