[WFLY-2335] fix BroadcastGroupAdd model population on JDK8
Running tests with JDK8 revealed an issue with that code: there is no
guarantee when populating the broadcast-group model that the connector
resources have already been populated and that we can check them.
That seems to be the case with JDK7 but not with JDK8 (likely due to
Collections implementation details).

Fix this by checking the validation of connector-refs in another MODEL
OperationStepHandler ran *after* all the other MODEL's OSHs.