WFLY-4219 Use Message Format Patterns
Most of the WildFly code uses message format patterns for debug logging
in order to avoid unnecessary allocations in normal operation. The
EJB3 subsystem client seems to be one of the few places where this is
not yet the case and debug messages are always created even if the
debug logging is disabled.

 * use `#debugf` for debug logging
 * use `#isDebugEnabled` for debug statements too complex for `#debugf`

Issue: WFLY-4219
https://issues.jboss.org/browse/WFLY-4219