[WFLY-4013] messaging's socket-binding are not registered
When the HornetQService is started (resp. stopped), the socket-bindings
used by the service are registered (resp. unregistered), so that their
runtime state is correct (calling :read-resource on them will display
them as bound).

Note that the socket-bindings referenced by the connectors are *not*
registered. We use them only to get a reference on the host/port of the
server but they are not opened (the client will open them).

JIRA: https://issues.jboss.org/browse/WFLY-4013