fix intermittent failures on JMSTopicManagementTestCase
Remove assertion that a consumer can not receive a message after the
topic is closed. There is a time window where the messaging client may
not have been advertised that the topic has been removed and the
consumer must be closed.
This assertion is not relevant to the tests. Ensuring that removing and
adding back the same topic removes all stored messages is enough