Added two tests: dump HTTPS request and request with upgrade to Websocket protocol
Added basic check of request dump values of request via HTTPS and HTTP with upgrade
to websocket protocol. This is just to ensure that dumping values are OK when used secured protocol.

To preserve readibility code is split in two files - in *TestCase there is server setup and tests definition
whereas in *TestImpl there is particular implementation of what is tested.